refactor: create new lsp package and move classes
